New Architecture: Modules/Restorables
This is a fairly significant architectural change.  The top-level model is now
something like this:  FMN has a list of modules, which support a getRestorables
method.  This method returns an array of restorables, which support a restore
method.  When the display is about to change, getRestorables is called on all
modules and the results are associated with the current display config.  After
the display changes, the array of restorables associated with the new display
config have "restore" sent to them.  This decouples the main app from the
concept of windows, making it easier to add non-window-related features.  For
example, we could let the user specify a shell script to be run on pre/post
display change events.

This patch also adds a key to the Info.plist file: ExcludedAppBundleIDs.  This
is an array of strings, each of which is the bundle identifier of an app to be
ignored.  At the moment, the identifier "com.apple.systempreferences" is on the
list, because it can cause big (~6 second) delays in display reconfiguration if
we try to handle it in certain circumstances.  Specifically, if I connect my PB
to an external monitor, set that to be the main screen, and then use System
Preferences to change the resolution of the internal LCD then I see big delays.

I really should have recorded these two changes separately, but now they're a bit intertwined.  I wish there was a way to split hunks...

darcs-hash:20060821214757-759cc-4b1dd4898c4130900f83a1b356ad321001ff1359.gz