Added Preference Pane
That's right boys and girls!  A new sucktastic preference pane!  But hey, it
works, provided FMN.app starts in under 1 second on your platform of choice.

Seriously, this is really awful.  It must be fixed.  The only thing that works
is the topmost toggle switch and it only works if the pane connects to the app
successfully.  There's a race condition when starting FMN.app.  The logic in the
pref pane is really feeble.

HOWEVAH, it *is* a preference pane, it *is* integrated into the project, it
*does* launch FMN.app, it *can* communicate with the app, and it *does* work if
you're gentle and kind-hearted.

I haven't yet disabled the GUI in FMN because it provides an easy way to quit the app while debugging.  Once the pref pane is solid that should be done.

darcs-hash:20060822052436-759cc-543062648dcbc2e80eb024786e99f31628021746.gz