Modules are now Bundles
The modularization that used to be weak is now strong.  The X11 and AX modules
are now self-contained bundles that are loaded at runtime.  This means that if
the user hasn't installed the X libraries it's no big deal -- the X11 plugin
just doesn't get loaded.

This program is now officially over-engineered.  ;^)  It is now possible to run
FMN and have it do *nothing*.  Just delete the PlugIns directory.

darcs-hash:20060823005157-759cc-c7a1df6ef5a16f3484df732b2bc298de580417f0.gz