Name consistency, and flower icon
This should probably be two patches, but that takes too much time.  Everything
should be called Forget-Me-Not now.  Also, I made an icon and used it for the
app/prefpane.  For some reason it won't show for the prefpane on my machine, but
oh well.  I did some fiddling with the flower to try to make it look a bit less
flat.  It works ok for the 32x32 icon, but it doesn't work as well for the
128x128 size.  I'm no artist.  The highlights are all in a layer in the flower-inkscape.svgz file, so they're easy to turn off.

Note: if you try to mess with the flower-inkscape file you should be aware that
there's a copy of the flower object in the highlights layer that's completely
transparent.  It's used as a clipping object for the highlights, because a
clipping object has to be above the object it clips.

darcs-hash:20060829190110-759cc-710cd4249f4ec477724a4ad45b774ceef519b420.gz