Accessibility API disabled warning
There's now a warning if the accessibility API is disabled.  It's ugly, with text instructions on how to enable the API.  It would be better to have a button that launches the correct pref pane automatically, but I'll leave that for another day.

darcs-hash:20060829212725-759cc-eb649df54690b6bff36eb55717254d011a876c59.gz