Prefpane Module Support
With this commit, we can now write modules that are to be loaded in the FMN preference pane. Right now, the way this works is the developer needs to write a module class that implements the FMNPrefpaneModule protocol, and then the resulting bundle must be placed in the Contents/PlugIns subdirectory of the Forget-Me-Not.prefpane bundle.

NOTE: As of this commit, the view specified by the module will not be displayed if the isTabControl member function returns 'YES'; this will probably be fixed in the near future (I just need to figure out where best the tab view should go).

darcs-hash:20060917014041-3181a-501805038a71d1de8d31236083e30aca74e3613a.gz