X11 Leopard Fixes
This resolves the problems we were having with the X server on Leopard.  It does
not cause the X server to be launched if it isn't already running but can
connect if it is running.  The changes should be backwards-compatible to Tiger.

darcs-hash:20071119232226-759cc-4b69750ad5a43eca4a7e92080844e984cbf2199f.gz