Hack for the (-40, Y) error
Leopard introduced a number of problems in the accessibility API, but one of the most annoying ones was that we randomly started getting bogus results for window position.  Luckily, it turns out it's not too hard to detect this -- the x coordinate is always -40 and the y coordinate is the height of a screen (don't know exactly which one yet) plus 22.  This patch just discards the info for any window that has x

darcs-hash:20071123235450-759cc-60cf8e4d94c9b9ade2b83fc133ac15913b46c603.gz