Get app list once instead of once per desktop
This is not just change for the sake of change.  It does actually appear to make
a difference in performance.  Sadly, it doesn't do what I hoped it would, which
was to prevent getting different AXWindow handles for the same window on
different desktops.

darcs-hash:20071124065055-759cc-52e8ee4a3aec2069b44083f9f3b2bd45cfeab84c.gz