Added an origin window
This change implements the origin window workaround for the bogosity in
Leopard display services.  The idea is that we put a small window at (0,0) and
measure all other window positions in relation to it.  That way, if the origin
moves to some weird place during the screen configuration change
pre-notification we'll still be ok, provided the windows retain their
relative positions, which does seem to be the case.

I'm removing the previous code that worked around the current particular
kind of brokenness (this should be much more general and robust) but I'm
leaving in the context apparatus since it doesn't add much overhead and it
might be useful in the future.

darcs-hash:20071124175307-759cc-ea0edc491b92d92dae80cea41630d7579cc911be.gz