An Origin Window That Works
It's funny.  I knew that the "origin as restorable" wouldn't work.  I did it the
right way.  Then I forgot that it wouldn't work and changed it to be the wrong
way.  Then it took me quite a while to figure out what was wrong...

<sigh>

So I've called the callback restoreFinished with the intention that there may
be restoreStarted, saveStarted, and saveFinished callbacks as well, but we can
add them when they become useful.

darcs-hash:20071125005210-759cc-da8d5ce288b23e286c189a5ad3608598acfd8227.gz